pyproject.toml
[tool.poetry]
name = "knowme"
version = "0.1.0"
description = ""
authors = ["abhinavkashyap <abhinav@comp.nus.edu.sg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
bs4 = "^0.0.2"
streamlit = "^1.35.0"
pypdf = "^4.2.0"
unstructured = {extras = ["pdf"], version = "^0.14.4"}
python-dotenv = "^1.0.1"
langgraph = "^0.0.62"


[tool.poetry.group.dev.dependencies]
faiss-cpu = "^1.8.0"
langchain = "^0.2.1"
ipykernel = "^6.29.4"
langchain-openai = "^0.1.8"
black = "^24.4.2"
langchain-community = "^0.2.1"
langchain-chroma = "^0.1.1"
langchainhub = "^0.1.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
